refactor: move AssessmentFormSchema to validator layer and add LoginSchema tests
Move AssessmentFormSchema and AssessmentFormData from the component into
src/lib/validators/assessments.ts, consistent with the project's validator
layer architecture. Update all three caller pages to import AssessmentFormData
from the validator module. Add unit tests for LoginSchema in
src/lib/validators/__tests__/users.test.ts.
