+++ src/lib/validators/assessments.ts
@@ -32,3 +32,26 @@ export const UpdateAssessmentSchema = AssessmentBaseSchema.partial()
 
 export type CreateAssessmentInput = z.infer<typeof CreateAssessmentSchema>;
 export type UpdateAssessmentInput = z.infer<typeof UpdateAssessmentSchema>;
+
+// Form-local schema: accepts YYYY-MM-DD strings (HTML date input format).
+// API-level date validation (ISO datetime) is handled server-side.
+export const AssessmentFormSchema = z
+  .object({
+    name: z.string().min(1, 'Name is required').max(200),
+    description: z.string().max(2000).optional(),
+    status: z.enum(['planning', 'in_progress', 'completed']),
+    test_date_start: z.string().optional(),
+    test_date_end: z.string().optional(),
+    project_id: z.string().optional(),
+  })
+  .refine(
+    (d) => {
+      if (d.test_date_start && d.test_date_end) {
+        return new Date(d.test_date_end) >= new Date(d.test_date_start);
+      }
+      return true;
+    },
+    { message: 'End date must be on or after start date', path: ['test_date_end'] }
+  );
+
+export type AssessmentFormData = z.infer<typeof AssessmentFormSchema>;
